# specs/004-identity/contracts/api.yaml
openapi: 3.0.3
info:
  title: RAGGit Workstation API (Single-Tenant)
  version: 1.3.0
  description: >
    Single-tenant on-prem RAG on AI Workstation (LanceDB local path +
    Ollama/LLamaSharp/ONNX). Desktop is thin client over LAN (no WAN at query
    time per FR-004/Constitution IV). 1.3.0 (004-identity, MINOR/ADDITIVE per
    FR-013): local per-person accounts — POST /api/auth/login and
    POST /api/auth/refresh issue workstation-signed JWT bearer tokens (8h);
    /api/users CRUD (Admin-only) manages the people directory; GET /api/auth/me
    gains ADDITIVE optional fields (displayName, username, sub) — 1.2.0
    required fields unchanged, clients MUST ignore unknown fields; Bearer
    security scheme added alongside retained ApiKey (machine/bootstrap per
    FR-012); credentials and session tokens traverse the LAN over HTTPS only
    (FR-008); 429 added for sign-in lockout (FR-009, 5 attempts / 15 min);
    Documents.createdBy and Queries attribution carry the stable person id for
    Local sessions while legacy "admin"/"employee" values remain valid
    (FR-007). No breaking changes: every 1.2.0 path, schema, and required
    field is preserved. 1.2.0 added xlsx MIME + rejection ordering; 1.1.0
    added GET /api/auth/me (extensible envelope) and pdf/docx 400 hardening.
servers:
  - url: https://ai-workstation.local:5001
    description: AI Workstation LAN — HTTPS required for credential and token traffic (FR-008)
  - url: http://localhost:5001
...
security:
  - ApiKey: []
  - Bearer: []
paths:
...
  /api/documents:
    post:
      summary: Upload document to singleton library (Admin only)
      description: >
        Chunks (512/50), embeds locally, indexes into LanceDB file. Rejects
        unsupported type/size per FR-006. Corrupted/invalid pdf/docx returns 400
        with no partial index (FR-006, 002 hardening). XLSX (1.2.0) deep
        validation, cell cap 413, no-extractable-content 400 — rejection
        ordering unchanged from 1.2.0. Dedupe by hash returns 200 with the
        existing Document. 1.3.0 semantics — when authenticated with a Local
        Bearer token, createdBy is set to the caller's stable person id
        (User.Id uuid, FR-007); ApiKey sessions keep "admin"/"employee".
      operationId: uploadDocument
      security:
        - ApiKey: []
        - Bearer: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file: { type: string, format: binary }
              required: [file]
      responses:
        '200':
          description: Duplicate — returns existing Document (same hash, not re-indexed)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        '201':
          description: Accepted for indexing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        '400':
          description: Bad request — corrupted/invalid content or unsupported type; no partial index
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '413':
          description: Payload too large — >100MB or >100000 spreadsheet cells
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    get:
      summary: List documents in singleton library
      operationId: listDocuments
      security:
        - ApiKey: []
        - Bearer: []
      responses:
        '200':
          description: Documents
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Document' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/documents/{id}:
    delete:
      summary: Delete document and purge its vectors (Admin only)
      operationId: deleteDocument
      security:
        - ApiKey: []
        - Bearer: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204': { description: Deleted, content: {} }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    ApiKey:
      type: apiKey
      in: header
      name: X-Api-Key
      description: >
        Single-tenant machine/bootstrap key (FR-012). Retained for machine
        integration and initial bring-up; NOT used for per-person attribution
        when a Local session exists (FR-007). Windows AD remains a documented
        future provider.
    Bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Workstation-signed HS256 JWT issued by POST /api/auth/login (004).
        8-hour lifetime; identifies a person (sub = stable person id), role,
        and display name. HTTPS-only transport (FR-008). Deactivated or
        locked-out accounts are refused per request (FR-010).
  schemas:
...
    Document:
      type: object
      properties:
        id: { type: string, format: uuid }
        filename: { type: string }
        mime: { type: string, enum: [application/pdf, application/vnd.openxmlformats-officedocument.wordprocessingml.document, application/vnd.openxmlformats-officedocument.spreadsheetml.sheet, text/plain, text/markdown] }
        size: { type: integer, format: int64 }
        hash: { type: string, description: SHA-256 hex for dedupe }
        status: { type: string, enum: [Indexing, Ready, Failed] }
        createdBy: { type: string, description: '1.3.0 — stable person id (User uuid) for Local sessions; legacy "admin"/"employee" for ApiKey sessions' }
        createdAt: { type: string, format: date-time }
      required: [id, filename, mime, size, hash, status, createdAt]
...
    Error:
      type: object
      properties:
        error: { type: string, example: unauthorized }
  responses:
...
    Unauthorized:
      description: Unauthorized — missing/invalid/expired credential, or inactive account (generic)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Forbidden — Employee cannot perform Admin-only actions (upload/delete/people management)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'